refactor: add createApp factories for services
