+++ prepchef/services/access-svc/src/tests/access.test.ts
@@ -1,12 +1,10 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import Fastify from 'fastify';
-import access from '../api/access';
+import { createApp } from '../index';
 import { messageBus } from '../adapters/messageBus';
 
 test('provisions access and emits event', async () => {
-  const app = Fastify();
-  await app.register(access);
+  const app = await createApp();
 
   const events: any[] = [];
   messageBus.once('access.provisioned', (payload) => events.push(payload));
@@ -24,8 +22,7 @@ test('provisions access and emits event', async () => {
 });
 
 test('revokes access and emits event', async () => {
-  const app = Fastify();
-  await app.register(access);
+  const app = await createApp();
 
   const events: any[] = [];
   messageBus.once('access.revoked', (payload) => events.push(payload));
