+++ prepchef/services/auth-svc/src/tests/auth.test.ts
@@ -1,18 +1,11 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import Fastify from 'fastify';
-import auth from '../api/auth';
+import { createApp } from '../index';
 
 const validCreds = { username: 'admin', password: 'secret' };
 
-async function buildApp() {
-  const app = Fastify();
-  await app.register(auth);
-  return app;
-}
-
 test('issues tokens for valid credentials', async () => {
-  const app = await buildApp();
+  const app = await createApp();
   const res = await app.inject({
     method: 'POST',
     url: '/auth/login',
@@ -27,7 +20,7 @@ test('issues tokens for valid credentials', async () => {
 });
 
 test('rejects invalid credentials', async () => {
-  const app = await buildApp();
+  const app = await createApp();
   const res = await app.inject({
     method: 'POST',
     url: '/auth/login',
@@ -37,7 +30,7 @@ test('rejects invalid credentials', async () => {
 });
 
 test('refreshes token with valid refresh token', async () => {
-  const app = await buildApp();
+  const app = await createApp();
   const loginRes = await app.inject({
     method: 'POST',
     url: '/auth/login',
@@ -56,7 +49,7 @@ test('refreshes token with valid refresh token', async () => {
 });
 
 test('rejects invalid refresh token', async () => {
-  const app = await buildApp();
+  const app = await createApp();
   const res = await app.inject({
     method: 'POST',
     url: '/auth/refresh',
@@ -66,7 +59,7 @@ test('rejects invalid refresh token', async () => {
 });
 
 test('cannot reuse refresh token after rotation', async () => {
-  const app = await buildApp();
+  const app = await createApp();
   const loginRes = await app.inject({
     method: 'POST',
     url: '/auth/login',
