+++ prepchef/services/notif-svc/src/tests/notifications.test.ts
@@ -1,39 +1,10 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import Fastify from 'fastify';
+import { createApp } from '../index';
 
 test('send notification', async () => {
-  const app = Fastify();
-  
-  const sentNotifications: any[] = [];
-  
-  app.post('/send', async (req, reply) => {
-    const { type, recipient_id, data } = req.body as any;
-    
-    if (!type || !recipient_id) {
-      return reply.code(400).send({
-        error: 'Missing required fields',
-        message: 'type and recipient_id are required'
-      });
-    }
-    
-    const notification = {
-      id: crypto.randomUUID(),
-      type,
-      recipient_id,
-      data: data || {},
-      sent_at: new Date().toISOString(),
-      channels: ['email', 'push']
-    };
-    
-    sentNotifications.push(notification);
-    
-    return reply.code(202).send({
-      notification_id: notification.id,
-      status: 'queued'
-    });
-  });
-  
+  const app = await createApp();
+
   const res = await app.inject({
     method: 'POST',
     url: '/send',
@@ -46,12 +17,10 @@ test('send notification', async () => {
       }
     }
   });
-  
+
   assert.equal(res.statusCode, 202);
   const body = res.json();
   assert.ok(body.notification_id);
   assert.equal(body.status, 'queued');
-  assert.equal(sentNotifications.length, 1);
-  assert.equal(sentNotifications[0].recipient_id, 'user-123');
 });
