+++ prepchef/services/payments-svc/src/tests/payments.test.ts
@@ -1,34 +1,9 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import Fastify from 'fastify';
+import { createApp } from '../index';
 
 test('create payment intent', async () => {
-  const app = Fastify();
-  
-  app.post('/intents', async (req, reply) => {
-    const { amount_cents, metadata } = req.body as any;
-    
-    if (!amount_cents || amount_cents <= 0) {
-      return reply.code(400).send({
-        error: 'Invalid amount',
-        message: 'Amount must be positive'
-      });
-    }
-    
-    // Mock Stripe payment intent creation
-    const paymentIntent = {
-      id: `pi_${crypto.randomUUID().replace(/-/g, '')}`,
-      amount: amount_cents,
-      currency: 'usd',
-      status: 'requires_payment_method',
-      client_secret: `pi_secret_${crypto.randomUUID().replace(/-/g, '')}`,
-      metadata: metadata || {},
-      created: Date.now()
-    };
-    
-    return reply.code(201).send(paymentIntent);
-  });
-  
+  const app = await createApp();
   const res = await app.inject({
     method: 'POST',
     url: '/intents',
@@ -40,7 +15,7 @@ test('create payment intent', async () => {
       }
     }
   });
-  
+
   assert.equal(res.statusCode, 201);
   const body = res.json();
   assert.ok(body.id.startsWith('pi_'));
@@ -50,21 +25,7 @@ test('create payment intent', async () => {
 });
 
 test('reject invalid payment amount', async () => {
-  const app = Fastify();
-  
-  app.post('/intents', async (req, reply) => {
-    const { amount_cents } = req.body as any;
-    
-    if (!amount_cents || amount_cents <= 0) {
-      return reply.code(400).send({
-        error: 'Invalid amount',
-        message: 'Amount must be positive'
-      });
-    }
-    
-    return reply.code(201).send({ id: 'pi_test' });
-  });
-  
+  const app = await createApp();
   const res = await app.inject({
     method: 'POST',
     url: '/intents',
@@ -72,7 +33,7 @@ test('reject invalid payment amount', async () => {
       amount_cents: -100
     }
   });
-  
+
   assert.equal(res.statusCode, 400);
   assert.ok(res.json().error.includes('Invalid'));
 });
