+++ prepchef/services/pricing-svc/src/tests/pricing.test.ts
@@ -1,35 +1,9 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import Fastify from 'fastify';
+import { createApp } from '../index';
 
 test('calculate pricing for booking', async () => {
-  const app = Fastify();
-  
-  app.post('/quote', async (req, reply) => {
-    const { listing_id, starts_at, ends_at } = req.body as any;
-    
-    // Calculate hours
-    const start = new Date(starts_at);
-    const end = new Date(ends_at);
-    const hours = (end.getTime() - start.getTime()) / (1000 * 60 * 60);
-    
-    // Mock pricing
-    const hourlyRate = 5000; // $50 in cents
-    const subtotal = Math.round(hours * hourlyRate);
-    const serviceFee = Math.round(subtotal * 0.15); // 15% platform fee
-    const tax = Math.round((subtotal + serviceFee) * 0.0875); // 8.75% tax
-    
-    return reply.send({
-      listing_id,
-      hours,
-      hourly_rate_cents: hourlyRate,
-      subtotal_cents: subtotal,
-      service_fee_cents: serviceFee,
-      tax_cents: tax,
-      total_cents: subtotal + serviceFee + tax
-    });
-  });
-  
+  const app = await createApp();
   const res = await app.inject({
     method: 'POST',
     url: '/quote',
@@ -39,7 +13,7 @@ test('calculate pricing for booking', async () => {
       ends_at: '2024-03-01T14:00:00Z'
     }
   });
-  
+
   assert.equal(res.statusCode, 200);
   const body = res.json();
   assert.equal(body.hours, 4);
@@ -49,33 +23,8 @@ test('calculate pricing for booking', async () => {
 });
 
 test('pricing validation for invalid duration', async () => {
-  const app = Fastify();
-  
-  app.post('/quote', async (req, reply) => {
-    const { starts_at, ends_at } = req.body as any;
-    
-    const start = new Date(starts_at);
-    const end = new Date(ends_at);
-    
-    if (end <= start) {
-      return reply.code(400).send({
-        error: 'Invalid duration',
-        message: 'End time must be after start time'
-      });
-    }
-    
-    const hours = (end.getTime() - start.getTime()) / (1000 * 60 * 60);
-    
-    if (hours < 2) {
-      return reply.code(400).send({
-        error: 'Minimum duration not met',
-        message: 'Minimum booking duration is 2 hours'
-      });
-    }
-    
-    return reply.send({ hours });
-  });
-  
+  const app = await createApp();
+
   // Test invalid duration
   const res1 = await app.inject({
     method: 'POST',
@@ -86,10 +35,10 @@ test('pricing validation for invalid duration', async () => {
       ends_at: '2024-03-01T09:00:00Z' // End before start
     }
   });
-  
+
   assert.equal(res1.statusCode, 400);
   assert.ok(res1.json().error.includes('Invalid'));
-  
+
   // Test minimum duration
   const res2 = await app.inject({
     method: 'POST',
@@ -100,7 +49,7 @@ test('pricing validation for invalid duration', async () => {
       ends_at: '2024-03-01T11:00:00Z' // Only 1 hour
     }
   });
-  
+
   assert.equal(res2.statusCode, 400);
   assert.ok(res2.json().message.includes('2 hours'));
 });
